# Optimise R-tree queries in the case of map matching