--- a/include/engine/geospatial_query.hpp
+++ b/include/engine/geospatial_query.hpp
@@ -88,13 +88,11 @@ template <typename RTreeT, typename DataFacadeT> class GeospatialQuery
                 return valid;
             });
         auto phantom_nodes = MakePhantomNodes(input_coordinate, results);
-        // there is no guarantee that `SearchInRange` will return results sorted by distance, 
+        // there is no guarantee that `SearchInRange` will return results sorted by distance,
         // but we may rely on it somewhere
-        std::sort(phantom_nodes.begin(), phantom_nodes.end(), [](const auto &lhs,
-                                                                        const auto &rhs)
-            {
-                return lhs.distance < rhs.distance;
-            });
+        std::sort(phantom_nodes.begin(),
+                  phantom_nodes.end(),
+                  [](const auto &lhs, const auto &rhs) { return lhs.distance < rhs.distance; });
         return phantom_nodes;
 #endif
     }