--- a/include/util/rectangle.hpp
+++ b/include/util/rectangle.hpp
@@ -1,6 +1,7 @@
 #ifndef OSRM_UTIL_RECTANGLE_HPP
 #define OSRM_UTIL_RECTANGLE_HPP
 
+#include "coordinate.hpp"
 #include "util/coordinate.hpp"
 #include "util/coordinate_calculation.hpp"
 
@@ -167,6 +168,26 @@ struct RectangleInt2D
                max_lon != FixedLongitude{std::numeric_limits<std::int32_t>::min()} &&
                min_lat != FixedLatitude{std::numeric_limits<std::int32_t>::max()} &&
                max_lat != FixedLatitude{std::numeric_limits<std::int32_t>::min()};
+    }
+
+    static double MetersPerLngDegree(const FixedLatitude lat)
+    {
+        constexpr static double kMetersPerDegreeLat = 110567.0f;
+
+        constexpr float kRadPerDeg = (3.14 /* TODO */ / 180.0f);
+
+        return std::cos(kRadPerDeg * static_cast<double>(toFloating(lat))) * kMetersPerDegreeLat;
+    }
+    static RectangleInt2D ExpandMeters(const Coordinate &coordinate, const double meters)
+    {
+        constexpr static double kMetersPerDegreeLat = 110567.0f;
+        const double lat_offset = meters / kMetersPerDegreeLat;
+        const double lon_offset = meters / MetersPerLngDegree(coordinate.lat);
+
+        return RectangleInt2D{coordinate.lon - toFixed(FloatLongitude{lon_offset}),
+                              coordinate.lon + toFixed(FloatLongitude{lon_offset}),
+                              coordinate.lat - toFixed(FloatLatitude{lat_offset}),
+                              coordinate.lat + toFixed(FloatLatitude{lat_offset})};
     }
 };
 } // namespace osrm::util