… unit_tests/util/static_rtree.cpp