# Issue

Since in map matching we don't need to find the N nearest projections, but simply all projections within a certain radius, instead of using a priority queue-based algorithm, we can simply find all edges intersecting the bounding box defined by this radius and then post-process the results to remove edges that are not within the required radius. According to our current simple map matching benchmark with a default search radius of 5m (which is effectively 15m in practice), this makes map matching up to 50% faster! However, with an increase in radius, this optimization becomes less noticeable since a larger radius increases the number of candidates, impacting the number of routes we need to compute in the map matching algorithm. But I have some ideas on how to optimize this as well in my future PRs.

Actual improvement can be seen below in the `match_ch` and `match_mld` benchmarks(the result in master corresponds to default/5m in this PR).



## Tasklist

 - [ ] CHANGELOG.md entry ([How to write a changelog entry](http://keepachangelog.com/en/1.0.0/#how))
 - [ ] update relevant [Wiki pages](https://github.com/Project-OSRM/osrm-backend/wiki)
 - [ ] add tests (see [testing documentation](https://github.com/Project-OSRM/osrm-backend/blob/master/docs/testing.md))
 - [ ] review
 - [ ] adjust for comments
 - [ ] cherry pick to release branch

## Requirements / Relations

 Link any requirements here. Other pull requests this PR is based on?


<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1101.99<br/>plain u32: 1093.37<br/>aliased double: 954.491<br/>plain double: 966.534 | aliased u32: 1098.31<br/>plain u32: 1101.72<br/>aliased double: 966.005<br/>plain double: 964.487 |
| json-render | String: 8.62683ms<br/>Stringstream: 11.6428ms<br/>Vector: 7.58827ms | String: 8.60096ms<br/>Stringstream: 11.7067ms<br/>Vector: 7.58147ms |
| match_ch | Default radius: <br/>8.23085ms/req at 82 coordinate<br/>0.100376ms/coordinate<br/>Radius 5m: <br/>8.19639ms/req at 82 coordinate<br/>0.099956ms/coordinate<br/>Radius 10m: <br/>19.3981ms/req at 82 coordinate<br/>0.236562ms/coordinate<br/>Radius 15m: <br/>41.5492ms/req at 82 coordinate<br/>0.506697ms/coordinate<br/>Radius 30m: <br/>320.866ms/req at 82 coordinate<br/>3.913ms/coordinate | Default radius: <br/>4.50758ms/req at 82 coordinate<br/>0.0549705ms/coordinate<br/>Radius 5m: <br/>4.48594ms/req at 82 coordinate<br/>0.0547065ms/coordinate<br/>Radius 10m: <br/>15.3305ms/req at 82 coordinate<br/>0.186957ms/coordinate<br/>Radius 15m: <br/>37.3413ms/req at 82 coordinate<br/>0.455381ms/coordinate<br/>Radius 30m: <br/>318.02ms/req at 82 coordinate<br/>3.87829ms/coordinate |
| match_mld | Default radius: <br/>7.43256ms/req at 82 coordinate<br/>0.090641ms/coordinate<br/>Radius 5m: <br/>7.12935ms/req at 82 coordinate<br/>0.0869433ms/coordinate<br/>Radius 10m: <br/>16.5298ms/req at 82 coordinate<br/>0.201583ms/coordinate<br/>Radius 15m: <br/>36.6781ms/req at 82 coordinate<br/>0.447294ms/coordinate<br/>Radius 30m: <br/>361.997ms/req at 82 coordinate<br/>4.4146ms/coordinate | Default radius: <br/>3.39512ms/req at 82 coordinate<br/>0.0414039ms/coordinate<br/>Radius 5m: <br/>3.37432ms/req at 82 coordinate<br/>0.0411503ms/coordinate<br/>Radius 10m: <br/>12.3558ms/req at 82 coordinate<br/>0.15068ms/coordinate<br/>Radius 15m: <br/>32.0192ms/req at 82 coordinate<br/>0.390477ms/coordinate<br/>Radius 30m: <br/>351.998ms/req at 82 coordinate<br/>4.29266ms/coordinate |
| packedvector | random write:<br/>std::vector 9792.03 ms<br/>util::packed_vector 82197.7 ms<br/>slowdown: 8.39435<br/>random read:<br/>std::vector 8512.14 ms<br/>util::packed_vector 33173.5 ms<br/>slowdown: 3.8972 | random write:<br/>std::vector 9835.57 ms<br/>util::packed_vector 74173.3 ms<br/>slowdown: 7.54133<br/>random read:<br/>std::vector 8557.96 ms<br/>util::packed_vector 29830.8 ms<br/>slowdown: 3.48573 |
| rtree | 1 result:<br/>208.575ms ->  0.0208575 ms/query<br/>10 results:<br/>243.318ms ->  0.0243318 ms/query | 1 result:<br/>220.757ms ->  0.0220757 ms/query<br/>10 results:<br/>242.402ms ->  0.0242402 ms/query |
<!-- BENCHMARK_RESULTS_END -->